requirements.txt
appnope==0.1.0
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
backcall==0.2.0
decorator==4.4.2
Django==3.0.8
django-cors-headers==3.4.0
django-extensions==3.0.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.6.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
wcwidth==0.2.5
wrapt==1.12.1
